Refactors data.py into multiple files
